Don't draw cached parent tile
Drawing the cached parent tile if the child is not available makes the UI very irresponsive in some situations. You might provoke this by  zooming into the map and scroll the map before the new tiles are loaded (I was using a TileDownloadLayer for this). In general the tile layer seems much more responsive without drawing cached parents. Give it a try if you don't believe it, you will notice the difference.