Add support for single-hgt-zips for hillshading (#1181)
rationale: the Sonny DEM (are there even others?) are currently distributed as single tile zips and you can conveniently get them nested in another zip. Unzipping that outer zip is easy, but all the tiles would be quite a lot of annoying work (or require scripting). As a bonus, on uncompressed filesystems the zipped tiles would save a lot of space and probably also performance (definitely also performance for tiles that are mostly ocean)